Warning fixes for win64
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@35921 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/dnd/dnd.cpp
@@ -522,7 +522,7 @@ public:
     virtual size_t GetFormatCount(Direction dir) const
     {
         // our custom format is supported by both GetData() and SetData()
-        size_t nFormats = 1;
+        ULONG nFormats = 1;
         if ( dir == Get )
         {
             // but the bitmap format(s) are only supported for output
